Added change check for CI deployment

+++ deploy.sh
@@ -25,12 +25,15 @@ head=$(git log --format="%h" -n 1)
 git checkout --quiet gh-pages
 cp -rf ../build/* .
 git add -A
-git status
 
-# Setup travis git user
-git config user.name "travis"
-git config user.email "travis@email.com"
+# Check for changes
+status=$(git status)
+echo "$status";
 
-# Commit and push
-git commit -m "CI Deployment to Github Pages ($user@$head)"
-git push --force --quiet "https://${GH_TOKEN}@$remote" gh-pages:gh-pages > /dev/null 2>&1
+# Setup travis git user, commit and push changes
+if [[ $status != *"nothing to commit"* ]] ; then
+  git config user.name "travis"
+  git config user.email "travis@email.com" 
+  git commit -m "CI Deployment to Github Pages ($user@$head)"
+  git push --force --quiet "https://${GH_TOKEN}@$remote" gh-pages:gh-pages > /dev/null 2>&1
+fi
